Fix for failing gameover test
The check_row_win method is not deep copying the @board array, so when
it is nil_filling the board, the actual board's arrays are being filled.

This is causing #full? to return true. The quick fix would be to reject
nil from #full? but that causes a lot of problems with the stacks, and
the better solution is to deep copy inside check_row_win so @board is not
mutated.